+++ mirage/scenarios/default.ts
@@ -17,7 +17,15 @@ const {
 } = config;
 
 export default function(server: Server) {
-    const userTraits = defaultLoggedOut ? [] : ['loggedIn', 'withInstitutions', 'withFiles'];
+    const userTraits = defaultLoggedOut ? [] :
+        [
+            'loggedIn',
+            'withFiles',
+            'withInstitutions',
+            'withSettings',
+            'withAlternateEmail',
+            'withUnconfirmedEmail',
+        ];
     const currentUser = server.create('user', ...userTraits);
 
     server.create('user-setting', { user: currentUser });
@@ -42,6 +50,8 @@ export default function(server: Server) {
         index: 0,
     });
 
+    const firstNode = server.create('node', {});
+    server.create('contributor', { node: firstNode, users: currentUser, index: 0 });
     const nodes = server.createList<Node>('node', 10, {
         currentUserPermissions: Object.values(Permission),
     }, 'withContributors');